Nidhi Projects/poet-webpage/backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-multipart==0.0.20
anthropic==0.40.0
PyPDF2==3.0.1
pdfplumber==0.11.4
python-docx==1.1.2
python-pptx==1.0.2
Pillow==11.0.0
python-dotenv==1.0.1
openpyxl==3.1.5

# Phase 1 — persistence layer
sqlalchemy==2.0.36
alembic==1.14.0
psycopg[binary]==3.2.3
pgvector==0.3.6
uuid-utils==0.10.0
